move proto generator to danube-core
