+++ danube-broker/Cargo.toml
@@ -31,7 +31,6 @@ metrics-exporter-prometheus = "0.15.3"
 serde_yaml = "0.9.34"
 
 [build-dependencies]
-tonic-build = "0.11"
 
 [lints]
 workspace = true
