+++ danube-client/Cargo.toml
@@ -13,6 +13,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+danube-core = {path = "../danube-core"}
 base64 = "0.22.1"
 futures = "0.3.30"
 futures-core = "0.3.30"
