+++ danube-client/src/dispatch_strategy.rs
@@ -1,4 +1,4 @@
-use crate::proto::{ReliableOptions as ProtoReliableOptions, TopicDispatchStrategy};
+use danube_core::proto::{ReliableOptions as ProtoReliableOptions, TopicDispatchStrategy};
 use serde::{Deserialize, Serialize};
 
 /// Dispatch strategy for a topic.
