+++ danube-client/src/lib.rs
@@ -2,10 +2,6 @@
 //!
 //! Danube-Client -- the Danube stream service client
 
-pub mod proto {
-    include!("proto/danube.rs");
-}
-
 mod client;
 pub use client::{DanubeClient, DanubeClientBuilder};
