+++ danube-core/Cargo.toml
@@ -5,6 +5,11 @@ edition = "2021"
 
 [dependencies]
 serde = { workspace = true }
+prost = { workspace = true }
+tonic = { workspace = true }
+
+[build-dependencies]
+tonic-build = "0.11"
 
 [lints]
 workspace = true
